Get rid of old proxy and accounts code
